+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.2.1] - 2023-07-12
 ### Fixed
 - Fix State EnteredTime and FinishedTime (#59)
 
@@ -28,7 +30,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/ManageIQ/floe/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/ManageIQ/floe/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/ManageIQ/floe/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/ManageIQ/floe/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/ManageIQ/floe/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/ManageIQ/floe/tree/v0.1.0